+++ doc/release_notes.md
@@ -16,6 +16,8 @@
     database error which is now returned instead of `String` for
     the [ErrorKind::DbError](crate::ErrorKind::DbError) enum variant.
 1.  Added support for fetching UROWID.
+1.  Added support for setting the session time zone from the environment
+    variable `ORA_SDTZ` or the client's local time zone (#9).
 1.  Eliminated hang when an error occurs during a DML returning statemnt
     ([issue 16](https://github.com/oracle/rust-oracledb/issues/16)).
 1.  Ensure that connections returned from a pool always start with a call